requirements-dev.txt
ruff
pandas-stubs